move zng serialization code out of package journal and into zngbytes
This commit moves the zng serialization code that was in package journal
into a new, top-level package called zngbytes.  There was no dependency
between the journal code and this serialization logic and after this
change, this functionality will be more obviously available without
reaching into package journal.
